requirements.txt
PyPDF2
python-pptx
pdf2image
reportlab
Pillow
comtypes